change callback to init32

+++ esphome/components/satellite1/microphone/__init__.py
@@ -36,7 +36,7 @@ CONF_PDM = "pdm"
 CONF_ON_PCM_DATA = "on_pcm_data"
 PCMDataTrigger = i2s_audio_ns.class_(
     "PCMDataTrigger",
-    automation.Trigger.template(cg.std_vector.template(cg.uint8).operator("ref")),
+    automation.Trigger.template(cg.std_vector.template(cg.int32).operator("ref")),
 )
